requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# LangGraph — agent orchestration
langgraph==0.2.55
langchain-core==0.3.15

# LangSmith — tracing and observability
langsmith==0.1.147

# GitHub API
PyGithub==2.4.0

# HTTP client (for posting PR comments async)
httpx==0.27.2

# Gemini LLM
google-genai==0.8.0

# Env management
python-dotenv==1.0.1